Move testing code snippets to a separate file
Can't remember if any of it will be useful later so don't want to delete any of
it, but currently it's cluttering the main file too much. I'll figure out what
to do with it whenever I get around to writing proper test, which should be done
in a separate file anyway.